rename translation component
