+++ packages/@coorpacademy-components/src/organism/content-translation/index.js
