+++ packages/@coorpacademy-components/src/organism/content-translation/style.css
