+++ packages/@coorpacademy-components/src/organism/content-translation/test/fixtures/default.js
