+++ packages/@coorpacademy-components/storybook/components.js
@@ -147,6 +147,7 @@ import OrganismBrandTable from './../src/organism/brand-table';
 import OrganismBrandUpload from './../src/organism/brand-upload';
 import OrganismCardsGrid from './../src/organism/cards-grid';
 import OrganismCart from './../src/organism/cart';
+import OrganismContentTranslation from './../src/organism/content-translation';
 import OrganismCourseSelection from './../src/organism/course-selection';
 import OrganismDiscussion from './../src/organism/discussion';
 import OrganismGetTheApp from './../src/organism/get-the-app';
@@ -158,7 +159,6 @@ import OrganismListItems from './../src/organism/list-items';
 import OrganismMoocImageSlider from './../src/organism/mooc/image-slider';
 import OrganismMoocFooter from './../src/organism/mooc-footer';
 import OrganismMoocHeader from './../src/organism/mooc-header';
-import OrganismPlaylistTranslation from './../src/organism/playlist-translation';
 import OrganismPopin from './../src/organism/popin';
 import OrganismResourceBrowser from './../src/organism/resource-browser';
 import OrganismSearchAndChipsResults from './../src/organism/search-and-chips-results';
@@ -791,6 +791,8 @@ import OrganismCardsGridFixtureEmpty from '../src/organism/cards-grid/test/fixtu
 import OrganismCardsGridFixtureLoading from '../src/organism/cards-grid/test/fixtures/loading';
 import OrganismCardsGridFixtureMany from '../src/organism/cards-grid/test/fixtures/many';
 import OrganismCartFixtureDefault from '../src/organism/cart/test/fixtures/default';
+import OrganismContentTranslationFixtureDefault from '../src/organism/content-translation/test/fixtures/default';
+import OrganismContentTranslationFixtureDelete from '../src/organism/content-translation/test/fixtures/delete';
 import OrganismCourseSelectionFixtureDefault from '../src/organism/course-selection/test/fixtures/default';
 import OrganismCourseSelectionFixtureEmpty from '../src/organism/course-selection/test/fixtures/empty';
 import OrganismCourseSelectionFixtureFiltered from '../src/organism/course-selection/test/fixtures/filtered';
@@ -831,8 +833,6 @@ import OrganismMoocHeaderFixtureOnHeaderActionHandlers from '../src/organism/moo
 import OrganismMoocHeaderFixtureTeamsFrench from '../src/organism/mooc-header/test/fixtures/teams-french';
 import OrganismMoocHeaderFixtureTeams from '../src/organism/mooc-header/test/fixtures/teams';
 import OrganismMoocHeaderFixtureUserChangePassword from '../src/organism/mooc-header/test/fixtures/user-change-password';
-import OrganismPlaylistTranslationFixtureDefault from '../src/organism/playlist-translation/test/fixtures/default';
-import OrganismPlaylistTranslationFixtureDelete from '../src/organism/playlist-translation/test/fixtures/delete';
 import OrganismPopinFixtureDefault from '../src/organism/popin/test/fixtures/default';
 import OrganismResourceBrowserFixtureArabic from '../src/organism/resource-browser/test/fixtures/arabic';
 import OrganismResourceBrowserFixtureAudio from '../src/organism/resource-browser/test/fixtures/audio';
@@ -1241,6 +1241,7 @@ export const components = {
     OrganismBrandUpload,
     OrganismCardsGrid,
     OrganismCart,
+    OrganismContentTranslation,
     OrganismCourseSelection,
     OrganismDiscussion,
     OrganismGetTheApp,
@@ -1251,7 +1252,6 @@ export const components = {
     OrganismListItems,
     OrganismMoocFooter,
     OrganismMoocHeader,
-    OrganismPlaylistTranslation,
     OrganismPopin,
     OrganismResourceBrowser,
     OrganismSearchAndChipsResults,
@@ -2216,6 +2216,10 @@ export const fixtures = {
     OrganismCart: {
       Default: OrganismCartFixtureDefault
     },
+    OrganismContentTranslation: {
+      Default: OrganismContentTranslationFixtureDefault,
+      Delete: OrganismContentTranslationFixtureDelete
+    },
     OrganismCourseSelection: {
       Default: OrganismCourseSelectionFixtureDefault,
       Empty: OrganismCourseSelectionFixtureEmpty,
@@ -2273,10 +2277,6 @@ export const fixtures = {
       TeamsFrench: OrganismMoocHeaderFixtureTeamsFrench,
       Teams: OrganismMoocHeaderFixtureTeams,
       UserChangePassword: OrganismMoocHeaderFixtureUserChangePassword
-    },
-    OrganismPlaylistTranslation: {
-      Default: OrganismPlaylistTranslationFixtureDefault,
-      Delete: OrganismPlaylistTranslationFixtureDelete
     },
     OrganismPopin: {
       Default: OrganismPopinFixtureDefault
